+++ spec/examples/recruiter_spec.rb
@@ -5,7 +5,7 @@ require './examples/recruiter'
 
 rast Recruiter do
   spec '#assess' do
-    prepare do |position, score|
+    prepare do |subject, position, score|
       allow(subject).to receive(:score) { score }
       allow(subject).to receive(:position) { position }
     end