+++ spec/examples/worker_spec.rb
@@ -5,7 +5,7 @@ require './examples/worker'
 
 rast Worker do
   spec '#goto_work?' do
-    prepare do |day_type, dow|
+    prepare do |subject, day_type, dow|
       allow(subject).to receive(:day_of_week) { dow.to_sym }
       allow(subject).to receive(:holiday?) { day_type == 'Holiday' }
     end