+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change log
 
+- 0.5.0.pre - Redesigned to have subject as first block parameter. Everything inside
+the prepare block now runs under the context of RSpec.
 - 0.4.2.pre - Fix 4 tokens bug, due to mis-configured converters
 - 0.4.1.pre - Fix negative? invocation.
 - 0.4.0.pre - Allow yaml-less configuration.